Added the diagonal transition after a day of  getting bugs and errors
Initially, I thought the easiest way would be to update the scene with the new animation, renamed the nodes so they made more sense and then replace the old LOCAL version on world.tscn to the new one. This caused a lot of errors as the new naming meant the code and variables could not be called upon, this ruining the launch of any levels. Attempts to redo the code by updaing the caribles being called on instead replaced the 'null' error with the hide() ad show() functions not being related to animation player nodes, which when replaced with visible.true/false which then made a new error with the those functions not being able to be used. At the end of the day I realised I could simply change the local transition animation player rather than update the root sceme and consequently have the variaables which call upon it break.